apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vnibb"
version = "0.1.0"
description = "Vietnam-First OpenBB Distribution - A headless, API-first financial analytics platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    { name = "VNIBB Team" }
]
keywords = ["vietnam", "stock", "finance", "openbb", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    # Core Framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Data Processing
    "pandas>=1.5.3",
    "openpyxl>=3.1.2",
    "numpy>=1.24.0",
    
    # Data Providers
    "vnstock>=3.0.0",
    # Premium packages (vnstock_data, vnstock_ta, vnstock_news, vnstock_pipeline)
    # are installed via CLI installer, not pip
    
    # Pipeline dependencies (for vnstock_pipeline)
    "duckdb>=1.2.0",
    "pyarrow>=14.0.1",
    
    # WebSocket support (for real-time streaming)
    "websockets>=12.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",  # PostgreSQL async driver (Supabase)
    "psycopg2-binary>=2.9.0",  # PostgreSQL sync driver
    "supabase>=2.0.0",  # Supabase client
    "python-jose[cryptography]>=3.3.0",  # JWT handling
    
    # Cache
    "redis>=5.0.0",
    
    # HTTP Client
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",
    
    # Scraping (fallback providers)
    "playwright>=1.41.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "orjson>=3.9.0",  # Fast JSON serialization
    
    # Scheduler (for automated syncs)
    "apscheduler>=3.10.0",
    
    # AI/LLM (for AI Copilot)
    "google-generativeai>=0.8.0",
    
    # Rate Limiting
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",  # For TestClient
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
scheduler = [
    "apscheduler>=3.10.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/vnibb/vnibb"
Documentation = "https://vnibb.dev/docs"
Repository = "https://github.com/vnibb/vnibb"

[project.scripts]
vnibb = "vnibb.cli:main"
vnibb-seed-stocks = "scripts.seed_stocks:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["vnibb*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]
